+++ server/common/server.ts
@@ -9,9 +9,6 @@ import * as cookieParser from 'cookie-parser';
 import * as helmet from 'helmet';
 import * as csrf from 'csurf';
 import * as Brakes from 'brakes';
-import { execute } from 'graphql';
-import { subscribe } from 'graphql/subscription';
-import { SubscriptionServer } from 'subscriptions-transport-ws';
 import { graphqlExpress, graphiqlExpress } from 'apollo-server-express';
 import * as middleware from 'swagger-express-middleware';
 import * as swaggerUi from 'swagger-ui-express';
@@ -139,6 +136,8 @@ export default class ExpressServer {
 
       const swaggerDocument = YAML.load('./server/common/swagger/Api.yaml');
       app.use('/swagger', swaggerUi.serve, swaggerUi.setup(swaggerDocument));
+
+
     });
 
 
@@ -148,39 +147,5 @@ export default class ExpressServer {
     return this.server;
   }
 
-  public listen(port: string = process.env.PORT): Application {
-    // tslint:disable
-    const welcome = (port) => console.log(`up and running in ${process.env.NODE_ENV || 'development'} @: ${os.hostname()} on port: ${port}`);
-    const ws = http.createServer(app)
-    ws.listen(port, () => {
-      welcome(port);
-    }
-    );
-
-    const subscriptionServer = new SubscriptionServer({
-      execute,
-      subscribe,
-      schema: myGraphQLSchema
-    }, {
-        server: ws,
-        path: '/subscriptions',
-      });
-
-    if (process.env.STREAM_HYSTRIX === 'true') {
-      const globalStats = Brakes.getGlobalStats();
-      http.createServer((req, res) => {
-        res.setHeader('Content-Type', 'text/event-stream;charset=UTF-8');
-        res.setHeader('Cache-Control', 'no-cache, no-store, max-age=0, must-revalidate');
-        res.setHeader('Pragma', 'no-cache');
-        globalStats.getHystrixStream().pipe(res);
-      }).listen(3001, () => {
-        console.log('---------------------');
-        console.log('Hystrix Stream now live at localhost:3001/hystrix.stream');
-        console.log('---------------------');
-      });
-    }
-
-    return app;
-  }
   // tslint:disable-next-line:eofline
 }
